Remove `ref` deletion from props
This is a leftover from a previous version of the library and doesn't look to be
needed.  Leaving it in place will cause a runtime error during development with
React v15 because `ref` is always defined on `props` as a "special" prop.
Trying to delete it will throw an error because props are frozen with
`Object.freeze`