Move React to a peer dependency + dev dependency
Moving React to a peer dependency plus a separate dev dependency for testing seems
to be the current standard practice for React libraries.  It allows for testing
a specific version of React while still outwardly supporting multiple versions.
Also, it helps to alleviate the multiple React problems with npm 3 users.