React 15 prep (#9)
* Remove `ref` deletion from props

This is a leftover from a previous version of the library and doesn't look to be
needed.  Leaving it in place will cause a runtime error during development with
React v15 because `ref` is always defined on `props` as a "special" prop.
Trying to delete it will throw an error because props are frozen with
`Object.freeze`

* Move React to a peer dependency + dev dependency

Moving React to a peer dependency plus a separate dev dependency for testing seems
to be the current standard practice for React libraries.  It allows for testing
a specific version of React while still outwardly supporting multiple versions.
Also, it helps to alleviate the multiple React problems with npm 3 users.